Get file uploads as e-mail attachments.
This does currently not use the sanitized filenames nor the stored
files. It simply grabs the blobs from the Symfony form.

To solve this more neatly, the FileUploadHandler should somehow have a
way of passing along its resulting array of file data back to the event.
Not sure if that's actually a best practice.

But hey, it works.